chore: project sidebar
