+++ frontend/src/layouts/ProjectLayout.vue
@@ -21,13 +21,6 @@
       :style="'INFO'"
       :title="$t('project.overview.info-slot-content')"
     />
-    <BBTabFilter
-      class="px-3 pb-2 border-b border-block-border"
-      :responsive="false"
-      :tab-item-list="tabItemList"
-      :selected-index="state.selectedIndex"
-      @select-index="selectTab"
-    />
   </HideInStandaloneMode>
 
   <div class="py-4 px-6">
@@ -40,18 +33,10 @@
 </template>
 
 <script lang="ts" setup>
-import { startCase } from "lodash-es";
-import { computed, nextTick, reactive, watch } from "vue";
-import { useI18n } from "vue-i18n";
-import { useRouter } from "vue-router";
-import { BBTabFilterItem } from "@/bbkit/types";
+import { computed } from "vue";
 import ArchiveBanner from "@/components/ArchiveBanner.vue";
 import HideInStandaloneMode from "@/components/misc/HideInStandaloneMode.vue";
-import {
-  useCurrentUserIamPolicy,
-  useCurrentUserV1,
-  useProjectV1Store,
-} from "@/store";
+import { useCurrentUserV1, useProjectV1Store } from "@/store";
 import { DEFAULT_PROJECT_V1_NAME } from "@/types";
 import { State } from "@/types/proto/v1/common";
 import { TenantMode } from "@/types/proto/v1/project_service";
@@ -60,15 +45,6 @@ import {
   hasWorkspacePermissionV1,
   hasPermissionInProjectV1,
 } from "@/utils";
-
-type ProjectTabItem = {
-  name: string;
-  hash: string;
-};
-
-interface LocalState {
-  selectedIndex: number;
-}
 
 const props = defineProps({
   projectSlug: {
@@ -81,16 +57,12 @@ const props = defineProps({
   },
 });
 
-const router = useRouter();
-const { t } = useI18n();
-
 const currentUserV1 = useCurrentUserV1();
 const projectV1Store = useProjectV1Store();
 
 const project = computed(() => {
   return projectV1Store.getProjectByUID(String(idFromSlug(props.projectSlug)));
 });
-const currentUserIamPolicy = useCurrentUserIamPolicy();
 
 const isDefaultProject = computed((): boolean => {
   return project.value.name === DEFAULT_PROJECT_V1_NAME;
@@ -98,75 +70,6 @@ const isDefaultProject = computed((): boolean => {
 
 const isTenantProject = computed((): boolean => {
   return project.value.tenantMode === TenantMode.TENANT_MODE_ENABLED;
-});
-
-const projectTabItemList = computed((): ProjectTabItem[] => {
-  if (!currentUserIamPolicy.isMemberOfProject(project.value.name)) {
-    return [{ name: t("common.databases"), hash: "databases" }];
-  }
-  if (
-    !currentUserIamPolicy.allowToChangeDatabaseOfProject(project.value.name)
-  ) {
-    const list = [{ name: t("common.databases"), hash: "databases" }];
-    if (!isDefaultProject.value) {
-      list.push(
-        ...[
-          { name: t("common.members"), hash: "members" },
-          { name: t("common.settings"), hash: "setting" },
-        ]
-      );
-    }
-    return list;
-  }
-
-  const list: (ProjectTabItem | null)[] = [
-    { name: t("common.overview"), hash: "overview" },
-    { name: t("common.branches"), hash: "branches" },
-    { name: t("common.databases"), hash: "databases" },
-
-    isTenantProject.value
-      ? { name: t("common.database-groups"), hash: "database-groups" }
-      : null,
-
-    isTenantProject.value
-      ? null // Hide "Change History" tab for tenant projects
-      : { name: t("common.change-history"), hash: "change-history" },
-
-    { name: startCase(t("slow-query.slow-queries")), hash: "slow-query" },
-
-    { name: t("common.activities"), hash: "activity" },
-    isDefaultProject.value
-      ? null
-      : { name: t("common.gitops"), hash: "gitops" },
-    isDefaultProject.value
-      ? null
-      : { name: t("common.webhooks"), hash: "webhook" },
-    isDefaultProject.value
-      ? null
-      : { name: t("common.members"), hash: "members" },
-    isDefaultProject.value
-      ? null
-      : { name: t("common.settings"), hash: "setting" },
-  ];
-  const filteredList = list.filter((item) => item !== null) as ProjectTabItem[];
-
-  return filteredList;
-});
-
-const findTabIndexByHash = (hash: string) => {
-  hash = hash.replace(/^#?/g, "");
-  const index = projectTabItemList.value.findIndex(
-    (item) => item.hash === hash
-  );
-  if (index >= 0) {
-    return index;
-  }
-  // otherwise fallback to the first tab
-  return 0;
-};
-
-const state = reactive<LocalState>({
-  selectedIndex: findTabIndexByHash(router.currentRoute.value.hash),
 });
 
 const allowEdit = computed(() => {
@@ -194,46 +97,4 @@ const allowEdit = computed(() => {
   }
   return false;
 });
-
-const tabItemList = computed((): BBTabFilterItem[] => {
-  return projectTabItemList.value.map((item) => {
-    return {
-      title: item.name,
-      alert: false,
-    };
-  });
-});
-
-const selectProjectTabOnHash = () => {
-  const { name, hash } = router.currentRoute.value;
-  if (name == "workspace.project.detail") {
-    const index = findTabIndexByHash(hash);
-    selectTab(index);
-  } else if (
-    name == "workspace.project.hook.create" ||
-    name == "workspace.project.hook.detail"
-  ) {
-    state.selectedIndex = findTabIndexByHash("webhook");
-  }
-};
-
-const selectTab = (index: number) => {
-  state.selectedIndex = index;
-  router.replace({
-    name: "workspace.project.detail",
-    hash: "#" + projectTabItemList.value[index].hash,
-  });
-};
-
-watch(
-  () => [router.currentRoute.value.hash],
-  () => {
-    nextTick(() => {
-      selectProjectTabOnHash();
-    });
-  },
-  {
-    immediate: true,
-  }
-);
 </script>
